revert back to original audio routing - further UI look development
with further testing this evening, I noticed my setup of audio routing was nearly impossible to manipulate in a clean way from the UI. reverting back to what we know works and continuing to push this. got fancy sliders and some other elements looking good which I am excited about.