Audio Device Refresh
seems to be a major issue across both mac and windows - touchdesigner does not have a good fix for this.
as it stands now, i have a button that you need to manually press to execute a script that refreshes the entire projects device data
need to test further and see how this will effect performance and if it will try and reset when any USB device is added/removed. would be annoying if it refreshed removing or adding a mouse, etcetera